requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
PyJWT
bcrypt
